# Repository_health_analyser/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RepoInput {
  /** Git repository URL (GitHub HTTPS URL) */
  url: string;
  /** Optional display name for the repo */
  name?: string;
  /**
     * Maximum commits to process (default 500)
     * @nullable
     */
  maxCommits?: number | null;
}

export type RepoStatusProperty = typeof RepoStatusProperty[keyof typeof RepoStatusProperty];


export const RepoStatusProperty = {
  pending: 'pending',
  ingesting: 'ingesting',
  ready: 'ready',
  error: 'error',
} as const;

export interface Repo {
  id: number;
  name: string;
  url: string;
  status: RepoStatusProperty;
  /** @nullable */
  totalCommits?: number | null;
  /** @nullable */
  processedCommits?: number | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  errorMessage?: string | null;
}

export type RepoStatusStatus = typeof RepoStatusStatus[keyof typeof RepoStatusStatus];


export const RepoStatusStatus = {
  pending: 'pending',
  ingesting: 'ingesting',
  ready: 'ready',
  error: 'error',
} as const;

export interface RepoStatus {
  repoId: number;
  status: RepoStatusStatus;
  /** @nullable */
  totalCommits?: number | null;
  /** @nullable */
  processedCommits?: number | null;
  /** @nullable */
  progressPercent?: number | null;
  /** @nullable */
  currentSha?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type RepoStatsDateRange = {
  /** @nullable */
  start?: string | null;
  /** @nullable */
  end?: string | null;
};

export interface RepoStats {
  repoId: number;
  totalCommits: number;
  analyzedCommits: number;
  avgComplexity: number;
  maxComplexity: number;
  avgCoupling: number;
  totalFiles: number;
  spikeCount: number;
  /** @nullable */
  mostChangedFile?: string | null;
  dateRange?: RepoStatsDateRange;
}

export interface CommitSummary {
  sha: string;
  shortSha: string;
  message: string;
  author: string;
  timestamp: string;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  coupling?: number | null;
  /** @nullable */
  churnRisk?: number | null;
  /** @nullable */
  filesChanged?: number | null;
  /** @nullable */
  linesAdded?: number | null;
  /** @nullable */
  linesRemoved?: number | null;
  isSpikeCommit: boolean;
}

export interface CommitPage {
  commits: CommitSummary[];
  total: number;
  limit: number;
  offset: number;
}

export interface FunctionMetric {
  name: string;
  file: string;
  complexity: number;
  /** @nullable */
  lineNumber?: number | null;
}

export interface CommitDetail {
  sha: string;
  shortSha: string;
  message: string;
  author: string;
  timestamp: string;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  coupling?: number | null;
  /** @nullable */
  churnRisk?: number | null;
  /** @nullable */
  filesChanged?: number | null;
  /** @nullable */
  linesAdded?: number | null;
  /** @nullable */
  linesRemoved?: number | null;
  isSpikeCommit: boolean;
  /** @nullable */
  complexityDelta?: number | null;
  /** @nullable */
  couplingDelta?: number | null;
  changedFiles: string[];
  topComplexFunctions: FunctionMetric[];
}

export interface TimelinePoint {
  sha: string;
  shortSha: string;
  timestamp: string;
  commitIndex: number;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  coupling?: number | null;
  /** @nullable */
  churnRisk?: number | null;
  /** @nullable */
  filesChanged?: number | null;
  isSpikeCommit: boolean;
  message: string;
  author: string;
}

export interface HotspotFile {
  filePath: string;
  churnCount: number;
  couplingScore: number;
  complexityScore: number;
  riskScore: number;
  /** @nullable */
  lastChangedSha?: string | null;
  /** @nullable */
  lastChangedAt?: string | null;
}

export type GraphNodeKind = typeof GraphNodeKind[keyof typeof GraphNodeKind];


export const GraphNodeKind = {
  file: 'file',
  function: 'function',
  class: 'class',
  module: 'module',
} as const;

export type GraphNodeChangeType = typeof GraphNodeChangeType[keyof typeof GraphNodeChangeType];


export const GraphNodeChangeType = {
  added: 'added',
  removed: 'removed',
  changed: 'changed',
  unchanged: 'unchanged',
} as const;

export interface GraphNode {
  id: string;
  label: string;
  kind: GraphNodeKind;
  /** @nullable */
  complexity?: number | null;
  /** @nullable */
  fanIn?: number | null;
  /** @nullable */
  fanOut?: number | null;
  changeType: GraphNodeChangeType;
}

export type GraphEdgeKind = typeof GraphEdgeKind[keyof typeof GraphEdgeKind];


export const GraphEdgeKind = {
  calls: 'calls',
  imports: 'imports',
  inherits: 'inherits',
  contains: 'contains',
} as const;

export type GraphEdgeChangeType = typeof GraphEdgeChangeType[keyof typeof GraphEdgeChangeType];


export const GraphEdgeChangeType = {
  added: 'added',
  removed: 'removed',
  unchanged: 'unchanged',
} as const;

export interface GraphEdge {
  source: string;
  target: string;
  kind: GraphEdgeKind;
  changeType: GraphEdgeChangeType;
}

export interface GraphDiff {
  sha: string;
  /** @nullable */
  parentSha?: string | null;
  nodes: GraphNode[];
  edges: GraphEdge[];
  addedNodes: number;
  removedNodes: number;
  changedNodes: number;
  addedEdges: number;
  removedEdges: number;
}

export interface SpikeExplanation {
  sha: string;
  /** @nullable */
  complexityBefore?: number | null;
  /** @nullable */
  complexityAfter?: number | null;
  /** @nullable */
  percentChange?: number | null;
  explanation: string;
  /** @nullable */
  isIntentional?: boolean | null;
  recommendations: string[];
  cached: boolean;
}

export interface ArchitecturalSummary {
  repoId: number;
  summary: string;
  keyTrends: string[];
  riskModules: string[];
  generatedAt: string;
  cached: boolean;
}

export interface RepoNarrative {
  repoId: number;
  narrative: string;
  highlights: string[];
  generatedAt: string;
  cached: boolean;
}