# Repository_health_analyser/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repos:
    get:
      operationId: listRepos
      tags: [repos]
      summary: List all ingested repositories
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repo"
    post:
      operationId: ingestRepo
      tags: [repos]
      summary: Ingest a new Git repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoInput"
      responses:
        "202":
          description: Ingestion started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "400":
          description: Invalid repository URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /repos/{repoId}/commits/{sha}/graph-diff:
    get:
      operationId: getGraphDiff
      tags: [graph]
      summary: Get knowledge graph diff for a commit (before vs after)
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Graph diff with added/removed/changed nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphDiff"
        "404":
          description: Commit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/commits/{sha}/spike-explanation:
    get:
      operationId: getSpikeExplanation
      tags: [insights]
      summary: Get LLM explanation for a complexity spike at this commit
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LLM-generated spike explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpikeExplanation"
        "404":
          description: Commit not found or no spike detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/summary:
    get:
      operationId: getArchitecturalSummary
      tags: [insights]
      summary: Get LLM architectural drift summary for the repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: LLM-generated architectural summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitecturalSummary"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repos/{repoId}/narrative:
    get:
      operationId: getRepoNarrative
      tags: [insights]
      summary: Get a full demo narrative of the repository health story
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: LLM-generated narrative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoNarrative"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RepoInput:
      type: object
      properties:
        url:
          type: string
          description: Git repository URL (GitHub HTTPS URL)
        name:
          type: string
          description: Optional display name for the repo
        maxCommits:
          type: ["integer", "null"]
          description: Maximum commits to process (default 500)
      required:
        - url
    Repo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, ingesting, ready, error]
        totalCommits:
          type: ["integer", "null"]
        processedCommits:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - name
        - url
        - status
        - createdAt
        - updatedAt
...
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        kind:
          type: string
          enum: [file, function, class, module]
        complexity:
          type: ["number", "null"]
        fanIn:
          type: ["integer", "null"]
        fanOut:
          type: ["integer", "null"]
        changeType:
          type: string
          enum: [added, removed, changed, unchanged]
      required:
        - id
        - label
        - kind
        - changeType
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        kind:
          type: string
          enum: [calls, imports, inherits, contains]
        changeType:
          type: string
          enum: [added, removed, unchanged]
      required:
        - source
        - target
        - kind
        - changeType
    GraphDiff:
      type: object
      properties:
        sha:
          type: string
        parentSha:
          type: ["string", "null"]
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        addedNodes:
          type: integer
        removedNodes:
          type: integer
        changedNodes:
          type: integer
        addedEdges:
          type: integer
        removedEdges:
          type: integer
      required:
        - sha
        - nodes
        - edges
        - addedNodes
        - removedNodes
        - changedNodes
        - addedEdges
        - removedEdges
    SpikeExplanation:
      type: object
      properties:
        sha:
          type: string
        complexityBefore:
          type: ["number", "null"]
        complexityAfter:
          type: ["number", "null"]
        percentChange:
          type: ["number", "null"]
        explanation:
          type: string
        isIntentional:
          type: ["boolean", "null"]
        recommendations:
          type: array
          items:
            type: string
        cached:
          type: boolean
      required:
        - sha
        - explanation
        - cached
        - recommendations
    ArchitecturalSummary:
      type: object
      properties:
        repoId:
          type: integer
        summary:
          type: string
        keyTrends:
          type: array
          items:
            type: string
        riskModules:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
        cached:
          type: boolean
      required:
        - repoId
        - summary
        - keyTrends
        - riskModules
        - generatedAt
        - cached
    RepoNarrative:
      type: object
      properties:
        repoId:
          type: integer
        narrative:
          type: string
        highlights:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
        cached:
          type: boolean
      required:
        - repoId
        - narrative
        - highlights
        - generatedAt
        - cached