# Repository_health_analyser/lib/api-zod/src/generated/types/architecturalSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RepoHealth API — git ingestion, code metrics, knowledge graphs, and LLM insights
 * OpenAPI spec version: 0.1.0
 */

export interface ArchitecturalSummary {
  repoId: number;
  summary: string;
  keyTrends: string[];
  riskModules: string[];
  generatedAt: Date;
  cached: boolean;
}